services/sentiment_analyzer/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
importlib-metadata==0.23
kafka-python==1.4.7
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.5
pytest==5.2.2
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
